+++ Application/Source/CameraVer2.h
@@ -15,7 +15,7 @@ public:
 	CameraVer2();
 	~CameraVer2();
 	void Init(Vector3 position = Vector3(0, 3, 8), const Vector3& view = Vector3(1, 0, 0), const Vector3& up = Vector3(0, 1, 0));
-	void Update(float x_offset, float y_offset);
+	void Update(double x_offset, double y_offset);
 	void Updatemovement(double dt);
 	void SetTarget(Vector3 target);
 	void SetPosition(Vector3 position);
