+++ Application/Source/Renderer.cpp
@@ -227,7 +227,7 @@ void Renderer::RenderTextOnScreen(Mesh* mesh, std::string text, Color color, flo
 	glEnable(GL_DEPTH_TEST);
 }
 
-void Renderer::RenderMeshOnScreen(Mesh* mesh, int x, int y, int sizex, int sizey)
+void Renderer::RenderMeshOnScreen(Mesh* mesh, float x, float y, float sizex, float sizey)
 {
 	glDisable(GL_DEPTH_TEST);
 	Mtx44 ortho;
